+++ pkgs/intl4x/lib/src/display_names/display_names_stub_4x.dart
@@ -1,0 +1,15 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import '../../display_names.dart';
+import '../data.dart';
+import '../locale/locale.dart';
+import 'display_names_impl.dart';
+
+DisplayNamesImpl getDisplayNames4X(
+  Locale locale,
+  Data data,
+  DisplayNamesOptions options,
+) =>
+    throw UnimplementedError('Cannot use ICU4X in web environments.');