+++ pkgs/intl4x/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add ICU4X support for number formatting.
 - Add resource identifier annotations.
 - Add ICU4X support for plural rules.
+- Add ICU4X support for display names.
 - Add ICU4X support for datetime formatting.
 
 ## 0.8.1